refactor: introduce github rest/graphql client lib
